REF: Move `resources`
skipci
