+++ octue/resources/__init__.py
@@ -1,8 +1,5 @@
-from .analysis import Analysis
-from .child import Child
 from .datafile import Datafile
 from .dataset import Dataset
 from .manifest import Manifest
 
-
-__all__ = "Analysis", "Child", "Datafile", "Dataset", "Manifest"
+__all__ = ("Datafile", "Dataset", "Manifest")
