+++ octue/templates/template-child-services/elevation_service/app.py
@@ -4,7 +4,7 @@ import random
 def run(analysis):
     """Run a mock analysis producing random integers as elevations to send to the parent.
 
-    :param octue.resources.Analysis analysis:
+    :param octue.twined.resources.Analysis analysis:
     :return None:
     """
     analysis.output_values = [random.randint(0, 5000) for location in analysis.input_values["locations"]]
