+++ octue/templates/template-child-services/wind_speed_service/app.py
@@ -4,7 +4,7 @@ import random
 def run(analysis):
     """Run a mock analysis that returns a random integer wind speed for each given location to send to the parent.
 
-    :param octue.resources.Analysis analysis:
+    :param octue.twined.resources.Analysis analysis:
     :return None:
     """
     analysis.output_values = [random.randint(0, 200) for location in analysis.input_values["locations"]]
