+++ octue/twined/resources/__init__.py
@@ -1,0 +1,4 @@
+from .analysis import Analysis
+from .child import Child
+
+__all__ = ("Analysis", "Child")
