+++ octue/twined/resources/child.py
@@ -4,8 +4,8 @@ import logging
 import os
 
 from octue.definitions import DEFAULT_MAXIMUM_HEARTBEAT_INTERVAL
-from octue.resources import service_backends
 from octue.twined.cloud.pub_sub.service import Service
+from octue.twined.resources import service_backends
 
 logger = logging.getLogger(__name__)
